Global Search (DB changed)
Changed db GLOBAL_SEARCH table and the default charset from utf8 to utf8mb4 to solve a problem when this had nothing to do with it after all. But since it is suggested everywhere I thought it to be a good thing so I kept it. Oh and global search works for both english and greek. TODO: add more entries in db for it